chore(heap): Remove redundant ElementsArray, rename SealableElementsArray to ElementsArray
